Testing the PCA results on data sets
Added code to test the kmeans performence in separation of  cluster for the whole data frame (having all columns) normalized with StandardScaler to the data frame which has been subjected to variance tresholding. After PCA and performing plothing of both clustering results that better separation of clusters is provided by the approach based on dimensionality reduction of the entire data set without  var_tresholding.