Added function measuring curvature of movment
This function uses the `splprep` and `splev` functions from the SciPy library to fit a spline to the trajectory data and evaluate the curvature of the trajectory at each point. The curvature is calculated as the inverse of the radius of the circle that best approximates the trajectory at each point. By convention, the sign of the curvature is positive for curves that bend counterclockwise (to the left) and negative for curves that bend clockwise (to the right).